+++ app/src/main/java/com/space/quizapp/common/extensions/ApiExtensions.kt
@@ -1,19 +1,10 @@
 package com.space.quizapp.common.extensions
 
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.flowWithLifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.recyclerview.widget.ListAdapter
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 import com.space.quizapp.common.resource.Result
-import com.space.quizapp.ui.home.QuizAdapter
-import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.launch
-import kotlin.coroutines.CoroutineContext
 
 
 fun <T : Any> Flow<T>.toResult(): Flow<Result<T>> =
@@ -22,16 +13,3 @@ fun <T : Any> Flow<T>.toResult(): Flow<Result<T>> =
     }.catch<Result<T>> { e ->
         emit(Result.Error(e))
     }.onStart { emit(Result.Loading) }
-
-fun <T : Any> Fragment.collectFlow(
-    flow: Flow<T>,
-    coroutineContext: CoroutineContext = Dispatchers.Unconfined,
-    lifecycleState: Lifecycle.State = Lifecycle.State.STARTED,
-    block: (T) -> Unit
-) {
-    viewLifecycleOwner.lifecycleScope.launch(coroutineContext) {
-        flow.flowWithLifecycle(lifecycle, lifecycleState).collect {
-            block(it)
-        }
-    }
-}
