+++ app/src/main/java/com/space/quizapp/common/extensions/CollectFlowExtensions.kt
@@ -1,0 +1,23 @@
+package com.space.quizapp.common.extensions
+
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.flowWithLifecycle
+import androidx.lifecycle.lifecycleScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.launch
+import kotlin.coroutines.CoroutineContext
+
+fun <T : Any> Fragment.collectFlow(
+    flow: Flow<T>,
+    coroutineContext: CoroutineContext = Dispatchers.Unconfined,
+    lifecycleState: Lifecycle.State = Lifecycle.State.STARTED,
+    block: (T) -> Unit
+) {
+    viewLifecycleOwner.lifecycleScope.launch(coroutineContext) {
+        flow.flowWithLifecycle(lifecycle, lifecycleState).collect {
+            block(it)
+        }
+    }
+}
