+++ app/src/main/java/com/space/quizapp/di/ApiModule.kt
@@ -33,4 +33,5 @@ object ApiModule {
                 ).build()
             )
         )
-        .build() }
+        .build()
+}
