+++ app/src/main/java/com/space/quizapp/domain/usecase/auth/AuthenticationUseCase.kt
@@ -12,5 +12,4 @@ class AuthenticationUseCase @Inject constructor(private val authenticationReposi
     suspend fun signUpUser(username: String) = authenticationRepository.signUpUser(username)
 
     fun getCurrentUserId() = authenticationRepository.getCurrentUserId()
-
 }
