+++ app/src/main/java/com/space/quizapp/presentation/model/QuizUIModel.kt
@@ -1,0 +1,7 @@
+package com.space.quizapp.presentation.model
+
+data class QuizUIModel(
+    val id: String,
+    val quizTitle: String,
+    val quizDescription: String,
+)
