+++ app/src/main/java/com/space/quizapp/presentation/model/UserUIModel.kt
@@ -1,0 +1,8 @@
+package com.space.quizapp.presentation.model
+
+
+data class UserUIModel(
+    val userId: String,
+    val username: String,
+    val userGPA: String,
+)
