--- app/src/main/java/com/space/quizapp/ui/base/BaseFragment.kt
@@ -1,47 +1,0 @@
-package com.space.quizapp.ui.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import androidx.viewbinding.ViewBinding
-import com.space.quizapp.common.types.Inflater
-import dagger.hilt.android.AndroidEntryPoint
-
-abstract class BaseFragment<VB : ViewBinding>(private val inflate: Inflater<VB>) : Fragment() {
-
-    private var _binding: VB? = null
-    val binding get() = _binding!!
-
-    abstract fun onBind()
-    open fun observes() {}
-    open fun listeners() {}
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = inflate.invoke(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        onBind()
-        observes()
-        listeners()
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-    protected fun toast(message: String) {
-        Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-    }
-}
