--- app/src/main/java/com/space/quizapp/ui/custom/AuthIllustrationBackground.kt
@@ -1,13 +1,0 @@
-package com.space.quizapp.ui.custom
-
-import android.content.Context
-import android.graphics.Canvas
-import android.util.AttributeSet
-import com.space.quizapp.ui.base.QuizBaseCustomView
-
-
-class AuthIllustrationBackground(context: Context, attrs: AttributeSet) :
-    QuizBaseCustomView(context, attrs) {
-    override fun onBind(canvas: Canvas) {
-    }
-}
