--- app/src/main/java/com/space/quizapp/ui/model/QuizUIModel.kt
@@ -1,7 +1,0 @@
-package com.space.quizapp.ui.model
-
-data class QuizUIModel(
-    val id: String,
-    val quizTitle: String,
-    val quizDescription: String,
-)
