--- app/src/main/java/com/space/quizapp/ui/model/UserUIModel.kt
@@ -1,8 +1,0 @@
-package com.space.quizapp.ui.model
-
-
-data class UserUIModel(
-    val userId: String,
-    val username: String,
-    val userGPA: String,
-)
